refactor: move providers to client
