+++ src/app/(client)/_components/_providers/provider.tsx
@@ -3,9 +3,9 @@ import type { ReactNode } from 'react';
 import { Provider as ReduxProvider } from 'react-redux';
 
 import { ErrorBoundary } from '@/app/(client)/_components/_providers/error-boundary';
-import { ThemeProvider } from '@/app/_providers/theme-context';
+import { ThemeProvider } from '@/app/(client)/_components/_providers/theme-context';
 
-import { store } from '../(client)/_store/store';
+import { store } from '../../_store/store';
 
 export default function Provider({ children }: { children: ReactNode }) {
   return (
